windows/SQL/requirements.txt
Flask>=2.0
mysql-connector-python>=8.0
pyinstaller>=5.0
pystray>=0.19
Pillow>=9.0
requests>=2.28.0
